pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipelineprobe"
version = "0.3.0"
description = "Instant Data Pipeline Audit Report for Airflow + dbt + modern warehouses"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "WillowVibe", email = "hello@willowvibe.com" }
]
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "jinja2>=3.1.2",
    "psycopg2-binary>=2.9.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "google-cloud-bigquery>=3.11.0",
    "snowflake-connector-python>=3.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
    "typer[all]>=0.9",
    "ruff>=0.4.0",
]

[project.scripts]
pipelineprobe = "pipelineprobe.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["pipelineprobe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
